# Make wrapdims more performant