Resolve performance issues associated with type instabilities in tables without types columns (e.g. DataFrames) by introducing a function barrier.