@willtebbutt 2022-08-31
MWE
```julia
using AxisKeys, BenchmarkTools, DataFrames, UniqueVectors

N1 = 10_000
N2 = 20_000
df1 = DataFrame(a=randn(N1), b=randn(N1), c=randn(N1))
df2 = DataFrame(a=randn(N2), b=randn(N2), c=randn(N2))
```

On master:
```julia
julia> @benchmark wrapdims($df1, UniqueVector, :a, :b, :c)
BenchmarkTools.Trial: 86 samples with 1 evaluation.
 Range (min … max):  39.477 ms … 319.292 ms  ┊ GC (min … max):  0.00% … 86.29%
 Time  (median):     57.984 ms               ┊ GC (median):    24.62%
 Time  (mean ± σ):   58.253 ms ±  40.063 ms  ┊ GC (mean ± σ):  27.76% ± 19.16%

  █    █                                                        
  █▁▅▅▁██▁▁▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅ ▁
  39.5 ms       Histogram: log(frequency) by time       288 ms <

 Memory estimate: 784.85 MiB, allocs estimate: 343473.

julia> @benchmark wrapdims($df2, UniqueVector, :a, :b, :c)
BenchmarkTools.Trial: 38 samples with 1 evaluation.
 Range (min … max):   99.133 ms … 375.354 ms  ┊ GC (min … max):  5.56% … 76.17%
 Time  (median):     121.219 ms               ┊ GC (median):    24.71%
 Time  (mean ± σ):   133.574 ms ±  55.636 ms  ┊ GC (mean ± σ):  31.79% ± 12.23%

      █▄                                                         
  ▅▅▁▁██▁▁▁▁▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅▁▁▁▅ ▁
  99.1 ms       Histogram: log(frequency) by time        375 ms <

 Memory estimate: 3.05 GiB, allocs estimate: 693483.
```

On this branch:
```julia
julia> @benchmark wrapdims($df1, UniqueVector, :a, :b, :c)
BenchmarkTools.Trial: 145 samples with 1 evaluation.
 Range (min … max):  22.657 ms … 296.069 ms  ┊ GC (min … max):  0.00% … 91.81%
 Time  (median):     37.247 ms               ┊ GC (median):    37.21%
 Time  (mean ± σ):   34.594 ms ±  30.208 ms  ┊ GC (mean ± σ):  31.32% ± 20.73%

  █  ▇▂                                                         
  █▄▁██▁▁▄▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄ ▄
  22.7 ms       Histogram: log(frequency) by time       263 ms <

 Memory estimate: 776.56 MiB, allocs estimate: 119.

julia> @benchmark wrapdims($df2, UniqueVector, :a, :b, :c)
BenchmarkTools.Trial: 60 samples with 1 evaluation.
 Range (min … max):  59.226 ms … 334.263 ms  ┊ GC (min … max): 14.27% … 84.38%
 Time  (median):     75.836 ms               ┊ GC (median):    31.16%
 Time  (mean ± σ):   84.400 ms ±  44.774 ms  ┊ GC (mean ± σ):  37.92% ± 10.61%

     █▇                                                         
  ▃▁▁██▁▃▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃ ▁
  59.2 ms         Histogram: frequency by time          308 ms <

 Memory estimate: 3.03 GiB, allocs estimate: 129.
```

Similarly, on master:
```julia
julia> @benchmark wrapdims($df1, :a, :b, :c)
BenchmarkTools.Trial: 23 samples with 1 evaluation.
 Range (min … max):  177.266 ms … 468.751 ms  ┊ GC (min … max):  0.00% … 59.44%
 Time  (median):     206.543 ms               ┊ GC (median):     8.63%
 Time  (mean ± σ):   221.506 ms ±  74.873 ms  ┊ GC (mean ± σ):  14.61% ± 15.87%

  ▆▁    ▁█                                                       
  ██▄▁▄▁██▄▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▄ ▁
  177 ms           Histogram: frequency by time          469 ms <

 Memory estimate: 783.56 MiB, allocs estimate: 343450.

julia> @benchmark wrapdims($df2, :a, :b, :c)
BenchmarkTools.Trial: 7 samples with 1 evaluation.
 Range (min … max):  655.493 ms … 929.792 ms  ┊ GC (min … max):  1.75% … 29.97%
 Time  (median):     675.576 ms               ┊ GC (median):     3.95%
 Time  (mean ± σ):   743.002 ms ± 120.580 ms  ┊ GC (mean ± σ):  12.59% ± 12.61%

  █   █     ▁                                            ▁    ▁  
  █▁▁▁█▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁█ ▁
  655 ms           Histogram: frequency by time          930 ms <

 Memory estimate: 3.04 GiB, allocs estimate: 693460.
```

while this branch obtains:
```julia
julia> @benchmark wrapdims($df1, :a, :b, :c)
BenchmarkTools.Trial: 30 samples with 1 evaluation.
 Range (min … max):  136.625 ms … 419.831 ms  ┊ GC (min … max):  0.00% … 64.49%
 Time  (median):     157.435 ms               ┊ GC (median):     5.70%
 Time  (mean ± σ):   170.681 ms ±  65.244 ms  ┊ GC (mean ± σ):  14.09% ± 15.57%

  █▁  ▃ █                                                        
  ██▄▁█▄█▁▁▁▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▄ ▁
  137 ms           Histogram: frequency by time          420 ms <

 Memory estimate: 775.88 MiB, allocs estimate: 96.

julia> @benchmark wrapdims($df2, :a, :b, :c)
BenchmarkTools.Trial: 9 samples with 1 evaluation.
 Range (min … max):  531.570 ms … 865.563 ms  ┊ GC (min … max):  1.38% … 33.35%
 Time  (median):     550.719 ms               ┊ GC (median):     3.78%
 Time  (mean ± σ):   612.509 ms ± 122.836 ms  ┊ GC (mean ± σ):  12.47% ± 12.91%

  ▁ ██▁     ▁                                   ▁             ▁  
  █▁███▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  532 ms           Histogram: frequency by time          866 ms <

 Memory estimate: 3.03 GiB, allocs estimate: 106.
```

In short, this PR substantially reduces allocations, which yields moderate performance improvements.

@rofinn any thoughts on where I should put some allocation-related unit tests?